Move the Tooltip component to the  directory
