+++ src/components/timeline/TrackMemoryGraph.js
@@ -15,7 +15,7 @@ import {
 } from '../../selectors/profile';
 import { getThreadSelectors } from '../../selectors/per-thread';
 import { ORANGE_50 } from 'photon-colors';
-import Tooltip from '../shared/Tooltip';
+import Tooltip from '../tooltip/Tooltip';
 import EmptyThreadIndicator from './EmptyThreadIndicator';
 import bisection from 'bisection';
